Move ha-input component
