+++ src/auth/ha-auth-form-string.ts
@@ -1,7 +1,7 @@
 import { customElement } from "lit/decorators";
 import { HaFormString } from "../components/ha-form/ha-form-string";
 import "../components/ha-icon-button";
-import "../components/ha-input";
+import "../components/input/ha-input";
 
 @customElement("ha-auth-form-string")
 export class HaAuthFormString extends HaFormString {
